Issue #465: Add job duration metric for monitor jobs * Update metrics names in Unit tests * Add comments in MonitoringTask * Remove explicit conversion to double

+++ metricshub-engine/src/main/java/org/sentrysoftware/metricshub/engine/strategy/AbstractStrategy.java
@@ -538,7 +538,7 @@ public abstract class AbstractStrategy implements IStrategy {
 		metricFactory.collectNumberMetric(
 			endpointHostMonitor,
 			jobDurationMetricKey,
-			(double) (endTime - startTime) / 1000, // Job duration in seconds
+			(endTime - startTime) / 1000.0, // Job duration in seconds
 			strategyTime
 		);
 	}
